Rename Extension and Tests files
